move consistency check from AssayProvider to AssayService
