+++ api/src/org/labkey/api/assay/AbstractAssayProvider.java
@@ -1736,64 +1736,4 @@ public abstract class AbstractAssayProvider implements AssayProvider
     {
         return supportsPlateMetadata() && Boolean.TRUE.equals(getBooleanProperty(protocol, PLATE_METADATA_PROPERTY_SUFFIX));
     }
-
-    @Override
-    public @NotNull Collection<Map<String, Object>> checkResults(Container container, User user, ExpProtocol protocol, ResultsCheckHelper helper)
-    {
-        if (helper != null)
-        {
-            SQLFragment sql = _generateSQL(container, user, protocol, helper);
-            if (sql != null)
-            {
-                return new SqlSelector(ExperimentService.get().getSchema().getScope(), sql).getMapCollection();
-            }
-            return Collections.emptyList();
-        }
-        else
-            throw new IllegalArgumentException("A ResultsCheckHelper must be provided");
-    }
-
-    @Override
-    public @NotNull <E> Collection<E> checkResults(Container container, User user, ExpProtocol protocol, ResultsCheckHelper helper, Class<E> clazz)
-    {
-        if (helper != null)
-        {
-            SQLFragment sql = _generateSQL(container, user, protocol, helper);
-            if (sql != null)
-            {
-                return new SqlSelector(ExperimentService.get().getSchema().getScope(), sql).getArrayList(clazz);
-            }
-            return Collections.emptyList();
-        }
-        else
-            throw new IllegalArgumentException("A ResultsCheckHelper must be provided");
-    }
-
-    private @Nullable SQLFragment _generateSQL(Container container, User user, ExpProtocol protocol, ResultsCheckHelper checker)
-    {
-        Logger log = checker.getLogger();
-        AssayProvider provider = AssayService.get().getProvider(protocol);
-        if (provider != null)
-        {
-            AssayProtocolSchema schema = provider.createProtocolSchema(user, container, protocol, null);
-            TableInfo assayDataTable = schema.createTable(AssayProtocolSchema.DATA_TABLE_NAME, checker.getContainerFilter());
-            if (assayDataTable != null)
-            {
-                List<ValidationError> errors = checker.isValid(protocol, assayDataTable);
-                if (!errors.isEmpty())
-                {
-                    errors.forEach(ve -> log.error(ve.getMessage()));
-                    return null;
-                }
-
-                return checker.getValidationSql(container, user, protocol, assayDataTable);
-            }
-            else
-                log.error("Assay data table not found for protocol : " + protocol.getName());
-        }
-        else
-            log.error("Assay provider not found for protocol : " + protocol.getName());
-
-        return null;
-    }
 }
